web/requirements.txt
Django==1.7.5
Markdown==2.6.1
django-cors-headers==1.0.0
django-filter==0.9.2
djangorestframework==3.1.0
djoser==0.2.1